+++ src/Spinorama/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 import { Box } from '@mui/material'
 
@@ -12,6 +14,20 @@ const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 	const settings: SpinoramaSettings = {
 		duration: duration || 10000
 	}
+
+	// Callbacks
+	const nextItem = React.useCallback(() => {}, [])
+
+	React.useEffect(() => {
+		// Call interval
+		const interval = setInterval(nextItem, settings.duration)
+
+		// Cleanup effect
+		return () => {
+			// Clear the interval when the component is unmounted
+			clearInterval(interval)
+		}
+	}, [nextItem, settings.duration])
 
 	return <Box className={`spinorama${className ? ` ${className}` : ''}`}>{children}</Box>
 }